New directory structore. Support for 64-bit build.

+++ source/plugin/scintilla.hpp
@@ -6,10 +6,10 @@ namespace scintilla {
 
     inline auto codepage( const HWND handle )
         -> int
-    { return ::SendMessage( handle, SCI_GETCODEPAGE, 0, 0 ); }
+    { return static_cast<int>( ::SendMessage( handle, SCI_GETCODEPAGE, 0, 0 ) ); }
 
     inline auto length( const HWND handle )
         -> int
-    { return ::SendMessage( handle, SCI_GETLENGTH, 0, 0 ); }
+    { return static_cast<int>( ::SendMessage( handle, SCI_GETLENGTH, 0, 0 ) ); }
 
 }  // namespace scintilla
